move movie use cases to movie package
