+++ domain/vod/src/main/java/details/usecase/movie/GetMovieCastUseCase.kt
@@ -1,4 +1,4 @@
-package details.usecase
+package details.usecase.movie
 
 import details.repository.MovieRepository
 import entity.Actor
@@ -7,5 +7,4 @@ class GetMovieCastUseCase(
     private val movieRepository: MovieRepository
 ) {
     suspend fun execute(id: Int): List<Actor> = movieRepository.getCast(id)
-
 }
