+++ domain/vod/src/main/java/details/usecase/movie/GetMovieDetailsUseCase.kt
@@ -1,4 +1,4 @@
-package details.usecase
+package details.usecase.movie
 
 import details.repository.MovieRepository
 import entity.Movie
@@ -7,5 +7,4 @@ class GetMovieDetailsUseCase(
     private val movieRepository: MovieRepository
 ) {
     suspend fun execute(id: Int): Movie = movieRepository.getMovieDetails(id)
-
 }
