+++ domain/vod/src/main/java/details/usecase/movie/GetMovieImagesUseCase.kt
@@ -1,4 +1,4 @@
-package details.usecase
+package details.usecase.movie
 
 import details.repository.MovieRepository
 
@@ -6,5 +6,4 @@ class GetMovieImagesUseCase(
     private val movieRepository: MovieRepository
 ) {
     suspend fun execute(id: Int): List<String> = movieRepository.getImages(id)
-
 }
